+++ include/stag/ED/ImageSmooth.h
