+++ include/stag/ED/ImageSmoothCV.h
