+++ include/stag/ED/LineSegment.h
