+++ include/stag/ED/MyMath.h
