+++ src/ED/ValidateEdgeSegments.cpp
@@ -3,7 +3,7 @@
 #include <memory.h>
 #include <math.h>
 
-#include "ValidateEdgeSegments.h"
+#include "stag/ED/ValidateEdgeSegments.h"
 
 #define MAX_GRAD_VALUE 128*256
